requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.1
    # via fastmcp
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.2
    # via uvicorn
cryptography==45.0.5
    # via authlib
cyclopts==3.22.5
    # via fastmcp
dnspython==2.7.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22
    # via rich-rst
email-validator==2.2.0
    # via pydantic
exceptiongroup==1.3.0
    # via fastmcp
fastmcp==2.11.0
    # via -r requirements.in
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.177.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
isodate==0.7.2
    # via openapi-core
jsonschema==4.25.0
    # via
    #   mcp
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2025.4.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.11.0
    # via openapi-spec-validator
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via werkzeug
mcp==1.12.3
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via openapi-core
openapi-core==0.19.5
    # via fastmcp
openapi-pydantic==0.5.1
    # via fastmcp
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.2
    # via openapi-core
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastmcp
    #   google-generativeai
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via rich
pyparsing==3.2.3
    # via httplib2
pyperclip==1.9.0
    # via fastmcp
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.2
    # via jsonschema-path
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   google-api-core
    #   jsonschema-path
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==14.1.0
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
rich-rst==1.3.1
    # via cyclopts
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via rfc3339-validator
sniffio==1.3.1
    # via anyio
sse-starlette==3.0.2
    # via mcp
starlette==0.47.2
    # via mcp
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.14.1
    # via
    #   google-generativeai
    #   openapi-core
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via mcp
werkzeug==3.1.1
    # via openapi-core
